change favicon to relative and https for script
change favicon.png to relative link so that it always work no matter the name/ip of the sites, and change the bootstrap.min.js to load via https instead of http. else if hosting the site on https, it will show as insecure because this scripts was being loaded via http.